pyproject.toml
[tool.poetry]
name = "gidra"
version = "0.1.0"
description = "Genshin Impact Proxy"
authors = ["Mero"]

[tool.poetry.dependencies]
python = "^3.10"
betterproto = "^2.0.0b4"
loguru = "^0.6.0"
bottle = "^0.12.21"
ec2b = {git = "https://github.com/GrownNed/ec2b.py"}
requests = "^2.28.1"
pycryptodome = "^3.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
